requirements.txt
flask
google-generativeai
python-dotenv
prometheus-client
requests
